# Optimizes parseIPv4