@AlexanderYastrebov 2020-12-31
This one https://github.com/inetaf/netaddr/pull/95#issuecomment-752800277 shows how brittle all such optimizations are