@josharian 2020-12-31
> This one #95 (comment) shows how brittle all such optimizations are

Sort of. But also not--it has a clear explanation (https://github.com/inetaf/netaddr/issues/102, which I just noticed independently). And some of the people who contribute to this repo know the compiler internals fairly well, and thus have a decent sense of what's likely to be brittle and what isn't. And if there is a regression in the compiler, we have the means to report it, fix it, and lock it in with a test.